Bug 34344: Make item types breadcrumbs uniform
This patch updates the breadcrumbs on the item types administration page
so that it reads "Item types" instead of "Item types administration,"
making it more consistent with other administration pages. The same
change is made to the page's main heading.

To test, apply the patch and go to Administration -> Item types.

The breadcrumbs should read "Home -> Administration -> Item types. The
page heading should read "Item types."

Signed-off-by: Caroline Cyr La Rose <caroline.cyr-la-rose@inlibro.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>