Bug 34382: Fix inconsistencies in Patron categories page titles, breadcrumbs, and header
This patch makes some changes to the patron categories administration
page so that the page title, breadcrumb navigation, and headings are
more consistent with each other.

To test, apply the patch and go to Administration -> Patron categories.

- Test all variations of the page:
  - Main page (listing categories)
  - New category
  - Edit category
  - Confirm deletion of category when the category is in use (there are
    existing patrons who have that category)
  - Confirm deletion of category when the category is not in use

Note: This patch contains whitespace changes, so please ignore
whitespace if you're checking the diff.

Signed-off-by: Amanda Hovey <amanda.hovey@parliament.nsw.gov.au>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>