Bug 34388: Fix inconsistencies in Patron restriction types page headers
This patch fixes a couple of inconsistencies in the style of the patron
restriction types administration screen, making sure the restriction
type name is consistently in single quotes in headings and formatting
the deletion confirmation page like other administration interfaces.

Note: This patch contains indentation spaces, so ignore whitespace when
viewing the diff.

To test, apply the patch and go to Administration -> Patron restriction
types.

- If necessary, add a custom restriction type.
- After saving it, edit it.
- The page title, breadcrumb, and page heading should match: "Modify
  restriction type 'X'

- From the list of restriction types, click "Delete" on one of your
  custom restriction types.
- On the deletion confirmation page you should see a standard
  alert-style dialog box and two button options: "Yes, delete" and "No,
  do not delete." Both should work correctly.

Signed-off-by: Caroline Cyr La Rose <caroline.cyr-la-rose@inlibro.com>

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>