Bug 34394: Fix inconsistencies in MARC Bibliographic framework test page title and breadcrumbs
This patch makes some changes to the MARC bibliographic framework test
page so that breadcrumbs, page title, and page heading are consistent
with each other.

The patch also creates standard message and alert dialogs to show the
status of the check. This is more consistent than having the message
only at the bottom of the table.

The patch also changes the term "itemnum" to "itemnumber."

To test, apply the patch and go to Admnistration -> MARC bibliographic
framework test.

- The page title, breadcrumb navigation, and page heading should match:
  "MARC bibliographic framework test."
- If your system has no errors, there should be a message-style dialog
  at the top of the page saying "Configuration okay."
- Create an error in your MARC configuration:
  - Go to Administration -> MARC frameworks -> MARC structure
  - Locate the 952 tag and click Actions -> View subfields
  - Edit subfield 9 and set the "Managed in tab" dropdown to something
    other than "ignore"
- Return to the framework test page. There should be an alert-style
  dialog at the top of the page, "Configuration errors found"

Signed-off-by: Caroline Cyr La Rose <caroline.cyr-la-rose@inlibro.com>

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>