Bug 34400: Inconsistencies in OAI sets page titles, breadcrumbs, and header
This patch fixes a couple of inconsistencies in the style of the
OIA sets administration screen, making sure the page title, breadcrumb
navigation, and page headers are consistent with each other.

The patch also corrects the structure of the form on the OAI set
mappings page so that the submit button is outside the "page-section"
area.

To test, apply the patch and go to Administration -> OAI sets. Test each
variation of the page:

- Main page
- New OAI set
- Edit OAI set
- OAI set mappings

Signed-off-by: Émily-Rose Francoeur <emily-rose.francoeur@inLibro.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>