Bug 34630: Update MARC21 cataloging plugin templates with consistent body class
This patch updates cataloging plugin templates to remove the inline
style "padding:1em" from the body tag. This style conflicts with fixed
Bootstrap footers.

A "value_builder" class is added to the body tag instead, the style
added to staff-global.scss. There shouldn't be any visible change to the
appearance of the pages.

To test, apply the patch and rebuild the staff interface CSS and clear
your browser cache if necessary.

- Go to Cataloging -> New record.
- In a default MARC21 installation you can test various plugin popups to
    confirm that they display correctly:
    - 000
    - 006
    - 007
    - 008

To visually confirm each affected template, the plugins can be loaded in
their own tab for testing the appearance of the page. I have a list of
direct links here:

https://gitlab.com/-/snippets/2555254
Signed-off-by: Émily-Rose Francoeur <emily-rose.francoeur@inLibro.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>