Bug 34380: Fix inconsistencies in Item types page titles, breadcrumbs, and header
This patch fixes inconsistencies in the item type administration page so
that page titles, breadcrumb navigation, and page headers are consistent
with each other and with other pages in Koha.

To test, apply the patch and go to Administration -> Item types. Check
each variation of the page to confirm that everything looks correct:

- Item types list
 - New item type
 - Edit item type
 - Confirm deletion of item type

Signed-off-by: Émily-Rose Francoeur <emily-rose.francoeur@inLibro.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>