Bug 34411: Inconsistencies in Additional fields page titles, breadcrumbs, and header
This patch makes changes to the additional fields administration
template in order to make it more consistent with other pages in Koha.

To test, apply the patch and go to Administration -> Additional fields.
View each variation of the page to confirm that breadcrumb navigation,
page title, and page headings are consistent with each other.

- Main list
  - Table view (e.g. aqbasket, aqorders, etc)
    - New field
    - Edit field

Signed-off-by: Salah Ghedda <salah.ghedda@inLibro.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>