Bug 34409: Fix inconsistencies in Audio alerts headers
This patch makes a minor change to the structure of the audio alerts
administration template so that the add/edit form is more consistent
with other similar forms.

The patch doesn't make any changes to page title and breadcrumbs because
the add/edit action doesn't trigger a page reload, so the required
change can't be made just with markup.

To test, apply the patch and enable the AudioAlerts system preference
if necessary.

- Go to Administration -> Audio alerts.
- Click "New alert" and confirm that the form looks correct.
- Edit an existing alert to confirm that this view also looks correct.

Signed-off-by: Salah Ghedda <salah.ghedda@inLibro.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>