Bug 34407: Inconsistencies in Z39.50 servers page titles, breadcrumbs, and header
This patch makes changes to the Z39.50 servers administration and
related templates in order to make them more consistent with other pages
in Koha.

One case has been added to page title and breadcrumb navigation: An
indication that a search has been performed. The subheading is modified
to make it clear that the search is a "starts with" search.

To test, apply the patch and go to Administration -> Z39.50/SRU
servers. View each variation of the page to confirm that breadcrumb
navigation, page title, and page headings are consistent with each
other.

- Z39.50/SRU servers list.
  - New Z39.50 server
  - New SRU server
  - Edit Z39.50 server
  - Edit SRU server
    - Modify SRU search field mapping for a bibliographic record
      server
    - Modify SRU search field mapping for an authority record server
  - Use the header search form to search for a server by name

Signed-off-by: Émily-Rose Francoeur <emily-rose.francoeur@inLibro.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>